Gate the arm-streaming trajex pipeline behind the repo's trajex build tags
The streaming package imported trajex unconditionally, breaking
CGO_ENABLED=0 no_cgo builds (e.g. linux/arm viam-server): trajex has no
buildable files without cgo.

Follow the components/arm/sim convention for trajex-backed code:
- coordinator.go, trajex_session.go, arm_stream_with_target_runway.go
  (and their tests, plus builtin's stream_test.go) build only under
  !windows && !no_cgo && viam_rdk_cgo_have_cxx20_rt.
- The pure-Go surface (StreamOptions, ParseStreamOptions,
  JointPositionsChItem) moves to an untagged options.go so the motion
  service's DoCommand layer compiles everywhere.
- coordinator_nocgo.go provides a fallback Run under the inverse tags
  that fails a session immediately with a clear "requires trajex" error.

Verified: CGO_ENABLED=0 GOOS=linux GOARCH=arm -tags no_cgo viam-server
build passes; tests pass with and without viam_rdk_cgo_have_cxx20_rt.
