+++ services/motion/builtin/stream_test.go
@@ -1,3 +1,5 @@
+//go:build !windows && !no_cgo && viam_rdk_cgo_have_cxx20_rt
+
 package builtin
 
 import (
